removed unused class
