+++ .env.example
@@ -1,7 +1,5 @@
 NEXT_PUBLIC_BASE_URL="http://localhost:3000"
 
-DEFAULT_STORE="swell"
-
 SWELL_STORE_ID=""
 SWELL_SECRET_KEY=""
 SWELL_PUBLIC_KEY=""
